code changes ver 0.8

+++ get_keyvault_secret.py
@@ -27,8 +27,8 @@ def welcome_menu():
 def main_menu():
     print("\nCreate a Keyvault [c] / Create KeyVault secret:value(s) [s]")
     print("List KeyVault secret names [ls] / Show secret/value pair [sh]")
-    print("Backup/restote Keyvault [br]")
     print("Delete Keyvault [dk] / delete secret [ds]")
+    print("Backup/restote Keyvault [br]")
     print("Quit program [q]")
 
 
@@ -47,10 +47,20 @@ def check_azure_keyvault(keyvaultname: str, rg: str):
         return False
 
 
+def azure_keyvault_set_create_secret_for_restor(keyvaultname: str, inputdict: dict):
+    for key, value in inputdict.items():
+        txt = "az keyvault secret set --vault-name {} -n {} --value {}"
+        cmd = txt.format(keyvaultname, key, value)
+        command = subprocess.run(["powershell", "-Command", cmd], capture_output=True)
+        if command.returncode != 0:
+            print("An error occurred: %s", command.stderr)
+        else:
+            print("SECRET " + key + " SUCCESSFULLY CREATE with " + value)
+
+
 def azure_keyvault_set_create_secret(keyvaultname: str):
     secretdict: dict = {}
-    secretcount = input("Please enter how many secrets you want to create: ")
-    # keyvaultname = input("Please type vault name: ")
+    secretcount = input("\nHow many secrets you want to create: ")
     for x in range(int(secretcount)):
         secretname = input("\nPlease type secret name: ")
         secretvalue = getpass.getpass("Enter secret value: ", stream=None)
@@ -66,24 +76,50 @@ def azure_keyvault_set_create_secret(keyvaultname: str):
             print("SECRET " + key + " SUCCESSFULLY CREATE with " + value)
 
 
-def azure_keyvault_delete_secret(keyvaultname: str):
-    secretlist = azure_keyvault_secret_list(keyvaultname)
-    print("following secrets exist in " + keyvaultname + "\n")
-    i = 1
-    for items in secretlist:
-        print(str(i) + ". " + items)
-        i = i + 1
-    deletesecretsamount = input("\nPlease enter how many secrets want to delete from " + keyvaultname + " keyvault: ")
-    for x in range(int(deletesecretsamount)):
-        secretname = input("Please enter secret name to delete: ")
-        txt = "az keyvault secret delete --vault-name {} -n {}"
-        cmd = txt.format(keyvaultname, secretname)
-        command = subprocess.run(["powershell", "-Command", cmd], capture_output=True)
-        if command.returncode != 0:
-            print("An error occurred: %s", command.stderr)
-        else:
-            print("SECRET " + secretname + " SUCCESSFULLY DELETED FROM " + keyvaultname)
-        
+def azure_keyvault_delete_secret(keyvaultname: str, secretlist: list):
+    # secretlist = azure_keyvault_secret_list(keyvaultname)
+    if len(secretlist) != 0:
+        print("following secrets exist in " + keyvaultname + "\n")
+        i = 1
+        for items in secretlist:
+            print(str(i) + ". " + items)
+            i = i + 1
+        deletesecretsamount = input("\nPlease enter how many secrets want to delete from " + keyvaultname
+                                    + " keyvault: ")
+        for x in range(int(deletesecretsamount)):
+            delete_permanent_question = input("Permanently delete the specified secret (yes/no): ")
+            if delete_permanent_question.casefold() == "no":
+                secretname = input("Please enter secret name for soft delete: ")
+                txt = "az keyvault secret delete --vault-name {} -n {}"
+                cmd = txt.format(keyvaultname, secretname)
+                command = subprocess.run(["powershell", "-Command", cmd], capture_output=True)
+                if command.returncode != 0:
+                    print("An error occurred: %s", command.stderr)
+                else:
+                    print("SECRET " + secretname + " SUCCESSFULLY DELETED FROM " + keyvaultname)
+            else:
+                secretname = input("Please enter secret name to delete permanently: ")
+                txt = "az keyvault secret delete --vault-name {} -n {}"
+                cmd = txt.format(keyvaultname, secretname)
+                command = subprocess.Popen(["powershell", "-Command", cmd], text=True)
+                while command.poll() is None:
+                    print("--Deleting secret--")
+                    sleep(5)
+                if command.returncode != 0:
+                    print("An error occurred: %s", command.stderr)
+                else:
+                    print("deleting secret " + secretname + " permanently")
+                    txt = "az keyvault secret purge --vault-name {} -n {}"
+                    cmd = txt.format(keyvaultname, secretname)
+                    command1 = subprocess.Popen(["powershell", "-Command", cmd], text=True)
+                    while command1.poll() is None:
+                        print("Please wait")
+                        sleep(10)
+                    else:
+                        print("SECRET " + secretname + " PERMANENTLY DELETED FROM " + keyvaultname + " SUCCESSFULLY")
+    else:
+        input("Press enter to continue")
+
 
 def azure_keyvault_secret_list(keyvaultname: str):
     entrylist = []
@@ -94,12 +130,7 @@ def azure_keyvault_secret_list(keyvaultname: str):
     for items in commandoutputlist:
         keyvaultentry = items.get("name")
         entrylist.append(keyvaultentry)
-    is_list_empty = len(entrylist)
-    print(is_list_empty)
-    if is_list_empty == 0:
-        return print("There is no secrets in " + keyvaultname + " keyvault")
-    else:
-        return entrylist
+    return entrylist
 
 
 def azure_keyvault_secret_show_by_secret_name(keyvaultname: str):
@@ -112,8 +143,21 @@ def azure_keyvault_secret_show_by_secret_name(keyvaultname: str):
     print("For SECRET NAME: " + secretname + "\nVALUE IS: " + secretvalue)
 
 
-def azure_keyvault_secret_show(keyvaultname: str):
-    inputlist = azure_keyvault_secret_list(keyvaultname)
+def azure_keyvault_secret_show_for_backup(keyvaultname: str, inputlist: list):
+    # inputlist = azure_keyvault_secret_list(keyvaultname)
+    keyvultsecretdict = {}
+    for item in inputlist:
+        txt = "az keyvault secret show --vault-name {} -n {}"
+        cmd = txt.format(keyvaultname, item)
+        commandoutput = subprocess.check_output(["powershell", "-Command", cmd], text=True)
+        jsonloader = json.loads(commandoutput)
+        secret_value = jsonloader["value"]
+        keyvultsecretdict[item] = secret_value
+    return keyvultsecretdict
+
+
+def azure_keyvault_secret_show(keyvaultname: str, inputlist: list):
+    # inputlist = azure_keyvault_secret_list(keyvaultname)
     keyvultsecretdict = {}
     i: int = 1
     for item in inputlist:
@@ -136,7 +180,7 @@ def azure_keyvault_create(keyvaultname: str, rg: str):
     command = subprocess.Popen(["powershell", "-Command", cmd], text=True)
     while command.poll() is None:
         print("--Creating--")
-        sleep(10)
+        sleep(15)
     if command.returncode != 0:
         print("An error occurred: %s", command.stderr)
     else:
@@ -151,7 +195,7 @@ def azure_keyvault_delete(keyvaultname: str, rg: str):
     command = subprocess.Popen(["powershell", "-Command", cmd], text=True)
     while command.poll() is None:
         print("--Deleting--")
-        sleep(10)
+        sleep(15)
     if command.returncode != 0:
         print("An error occurred: %s", command.stderr)
     else:
@@ -181,7 +225,6 @@ if __name__ == '__main__':
             source_keyvaultname = input("Please type vault name: ")
             print("Checking if " + source_keyvaultname + " already exist in " + source_rg + " Resource group")
             is_keyvault_exist = check_azure_keyvault(source_keyvaultname, source_rg)
-            print(is_keyvault_exist)
             if is_keyvault_exist:
                 print("\nKeyvault " + source_keyvaultname + " already exist")
                 input("Press enter to return main menu: ")
@@ -191,18 +234,31 @@ if __name__ == '__main__':
                 input("\n Press enter to return main menu: ")
                 trigger = False
 
+        # Create Keyvault secret/value pair block
         elif answer == "s":
-            if not source_keyvaultname:
-                source_keyvaultname = input("Please type vault name: ")
-                azure_keyvault_set_create_secret(source_keyvaultname)
-                input("\nPress enter to return main menu: ")
-                trigger = False
+            source_rg = input("\nPlease type RG name (default is " + default_rg + "): ") or default_rg
+            source_keyvaultname = input("Please type vault name: ")
+            print("Checking if " + source_keyvaultname + " already exist in " + source_rg + " Resource group")
+            is_keyvault_exist = check_azure_keyvault(source_keyvaultname, source_rg)
+            if is_keyvault_exist:
+                keyvault_list_s = azure_keyvault_secret_list(source_keyvaultname)
+                if len(keyvault_list_s) != 0:
+                    item_count = 1
+                    print("Following secrets exist in " + source_keyvaultname + " keyvault")
+                    for items in keyvault_list_s:
+                        print(str(item_count) + ". " + items)
+                        item_count = item_count + 1
+                    azure_keyvault_set_create_secret(source_keyvaultname)
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
+                else:
+                    print("There is no secrets in " + source_keyvaultname + " keyvault")
+                    azure_keyvault_set_create_secret(source_keyvaultname)
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
             else:
-                source_keyvaultname = input("Please type vault name " +
-                                            "(or press enter to use current keyvault " + source_keyvaultname + " ): ") \
-                                      or source_keyvaultname
-                azure_keyvault_set_create_secret(source_keyvaultname)
-                input("\nPress enter to return main menu: ")
+                print("\nThere is no such keyvault")
+                input("Press enter to return main menu: ")
                 trigger = False
 
         elif answer == "ls":
@@ -211,17 +267,22 @@ if __name__ == '__main__':
             print("Checking if " + source_keyvaultname + " already exist in " + source_rg + " Resource group")
             is_keyvault_exist = check_azure_keyvault(source_keyvaultname, source_rg)
             if is_keyvault_exist:
-                item_count: int = 1
-                keyvault_list = azure_keyvault_secret_list(source_keyvaultname)
-                print("Following secrets exist in " + source_keyvaultname + "\n")
-                for items in keyvault_list:
-                    print(str(item_count) + ". " + items)
-                    item_count = item_count + 1
-                input("\nPress enter to return main menu: ")
-                trigger = False
+                secret_list_ls = azure_keyvault_secret_list(source_keyvaultname)
+                if len(secret_list_ls) != 0:
+                    item_count: int = 1
+                    print("Following secrets exist in " + source_keyvaultname + "\n")
+                    for items in secret_list_ls:
+                        print(str(item_count) + ". " + items)
+                        item_count = item_count + 1
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
+                else:
+                    print("There is no secrets in " + source_keyvaultname + " keyvault")
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
             else:
-                print("\nThere is no such keyvault")
-                input("Press enter to return main menu: ")
+                print("There is no such keyvault")
+                input("\nPress enter to return main menu: ")
                 trigger = False
 
         elif answer == "sh":
@@ -230,12 +291,18 @@ if __name__ == '__main__':
             print("Checking if " + source_keyvaultname + " already exist in " + source_rg + " Resource group")
             is_keyvault_exist = check_azure_keyvault(source_keyvaultname, source_rg)
             if is_keyvault_exist:
-                azure_keyvault_secret_show(source_keyvaultname)
-                input("\nPress enter to return main menu: ")
-                trigger = False
+                secret_list_sh = azure_keyvault_secret_list(source_keyvaultname)
+                if len(secret_list_sh) != 0:
+                    azure_keyvault_secret_show(source_keyvaultname, secret_list_sh)
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
+                else:
+                    print("There is no secrets in " + source_keyvaultname + " keyvault")
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
             else:
-                print("\nThere is no such keyvault")
-                input("Press enter to return main menu: ")
+                print("There is no such keyvault")
+                input("\nPress enter to return main menu: ")
                 trigger = False
 
         elif answer == "dk":
@@ -244,9 +311,22 @@ if __name__ == '__main__':
             print("Checking if " + source_keyvaultname + " already exist in " + source_rg + " Resource group")
             is_keyvault_exist = check_azure_keyvault(source_keyvaultname, source_rg)
             if is_keyvault_exist:
-                azure_keyvault_delete(source_keyvaultname, source_rg)
-                input("\nPress enter to return main menu: ")
-                trigger = False
+                secret_list_dk = azure_keyvault_secret_list(source_keyvaultname)
+                if len(secret_list_dk) == 0:
+                    azure_keyvault_delete(source_keyvaultname, source_rg)
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
+                else:
+                    print("There are " + str(len(secret_list_dk)) + " secrets exist in " +
+                          source_keyvaultname + " keyvault")
+                    delete_answer = input("Are you sure to delete (yes/no): ")
+                    if delete_answer.casefold() == "yes":
+                        azure_keyvault_delete(source_keyvaultname, source_rg)
+                        input("\nPress enter to return main menu: ")
+                        trigger = False
+                    else:
+                        input("\nPress enter to return main menu: ")
+                        trigger = False
             else:
                 print("\nThere is no such keyvault")
                 input("Press enter to return main menu: ")
@@ -258,12 +338,68 @@ if __name__ == '__main__':
             print("Checking if " + source_keyvaultname + " already exist in " + source_rg + " Resource group")
             is_keyvault_exist = check_azure_keyvault(source_keyvaultname, source_rg)
             if is_keyvault_exist:
-                azure_keyvault_delete_secret(source_keyvaultname)
-                input("\nPress enter to return main menu: ")
-                trigger = False
+                secret_list_ds = azure_keyvault_secret_list(source_keyvaultname)
+                if len(secret_list_ds) != 0:
+                    azure_keyvault_delete_secret(source_keyvaultname, secret_list_ds)
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
+                else:
+                    print("There is no secrets in " + source_keyvaultname + " keyvault to delete")
+                    input("\nPress enter to return main menu: ")
+                    trigger = False
             else:
                 print("\nThere is no such keyvault")
                 input("Press enter to return main menu: ")
+                trigger = False
+
+        elif answer == "br":
+            source_rg = input("\nPlease type source RG name (default is " + default_rg + "): ") or default_rg
+            source_keyvaultname = input("Please type source vault name: ")
+            print("Checking if " + source_keyvaultname + " already exist in " + source_rg + " Resource group")
+            is_source_keyvault_exist = check_azure_keyvault(source_keyvaultname, source_rg)
+
+            if is_source_keyvault_exist:
+                print("Keyvault " + source_keyvaultname + " exist in " + source_rg + " resource group")
+                sleep(1)
+                print("Retriving secrets from " + source_keyvaultname + " keyvault")
+                source_keyvaultname_secret_list = azure_keyvault_secret_list(source_keyvaultname)
+
+                if len(source_keyvaultname_secret_list) != 0:
+                    source_keyvaultname_dict = azure_keyvault_secret_show_for_backup(source_keyvaultname, source_keyvaultname_secret_list)
+                    print("Backup completed")
+                    sleep(1)
+                    print("Starting resore process")
+                    destination_rg = input("Please type destination RG name (default is " + default_rg + "): ") or \
+                                     default_rg
+                    destination_keyvaultname = input("Please type destination vault name: ")
+                    print("Checking if " + destination_keyvaultname + " keyvault already exist in destination" +
+                          destination_rg + " Resource group")
+                    is_destination_keyvault_exist = check_azure_keyvault(destination_keyvaultname, destination_rg)
+
+                    if not is_destination_keyvault_exist:
+                        print("destination " + destination_keyvaultname + " doesn't exist in destination" +
+                              destination_rg + " resource group")
+                        sleep(1)
+                        print("Creating destination " + destination_keyvaultname + " keyvault in destination " +
+                              destination_rg + " resource group")
+                        azure_keyvault_create(destination_keyvaultname, destination_rg)
+                        input("Press enter to start restore to destination " + destination_keyvaultname + " keyvault")
+                        print("Creating secret/value for destination " + destination_keyvaultname + " Keyvault")
+                        azure_keyvault_set_create_secret_for_restor(destination_keyvaultname, source_keyvaultname_dict)
+                    else:
+                        print("Destination " + destination_keyvaultname + " keyvault alerady exist")
+                        input("Press enter to start restoration to destination " + destination_keyvaultname + " keyvault: ")
+                        print("Creating secret/value for destination " + destination_keyvaultname + " Keyvault")
+                        azure_keyvault_set_create_secret_for_restor(destination_keyvaultname, source_keyvaultname_dict)
+                else:
+                    print("There is no secrets in " + source_keyvaultname + " keyvault to backup")
+                    print("Back to main menu and create secrets first")
+                    input("Press enter to return main menu: ")
+                    trigger = False
+            else:
+                print("Source " + source_keyvaultname + " keyvault doesn't exist")
+                print("Back to main menu and create keyvault first")
+                input("\nPress enter to return main menu: ")
                 trigger = False
 
         elif answer == "q":
